# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plateforme E-learning",
        "description": "Documentation API pour la plateforme e-learning",
        "contact": {
            "email": "diopawa425@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API Local"
        }
    ],
    "paths": {
...
        "/api/courses/{course}/quizzes": {
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Lister tous les quiz",
                "operationId": "98aa57513947a36ab3485b4a645d8ad8",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des quiz",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Quiz"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Création de nouvelle quiz",
                "operationId": "0356317b77518a698e70bac4e788c672",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Quiz"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Leçon créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Quiz"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/courses/{course}/quizzes/submit": {
            "post": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Soumettre un quiz avec les réponses de l'utilisateur",
                "description": "Soumet les réponses de l'utilisateur à un quiz, enregistre les résultats, et calcule le score.",
                "operationId": "submitQuiz",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quiz_id",
                                    "answers"
                                ],
                                "properties": {
                                    "quiz_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "answer_ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 42
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quiz soumis avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Quiz soumis avec succès"
                                        },
                                        "score": {
                                            "type": "string",
                                            "example": "3/5"
                                        },
                                        "percentage": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 60
                                        },
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "question_id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "correct_answer_ids": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "selected_answer_ids": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "is_correct": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Déjà effectué"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Quiz non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Answer": {
                "properties": {
                    "text": {
                        "type": "string",
                        "example": "Contenu reponse"
                    },
                    "is_correct": {
                        "type": "boolean",
                        "example": "true",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Question": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "multiple_choice",
                        "nullable": true
                    },
                    "text": {
                        "type": "string",
                        "example": "Contenu quiz"
                    },
                    "question": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Answer"
                        }
                    }
                },
                "type": "object"
            },
            "Quiz": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Quiz 1"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Entrez le token Sanctum reçu après connexion.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}